Remove Flow wrapper classes and adjust tests
Discussion about restrucuring the Flow shows that at least
the early participants in the API discussion favour a nested
structure over flat derived classes. Wrapper classes were
suggested to make clear that no investment into a non-convex
Flow was possible. As it now is, no reason seems to remain to
have them.

Related issues:
* Resturecure flow (#556)
* InvestmentFlow wrapper (#799)
