+++ src/oemof/solph/flows/__init__.py
@@ -13,14 +13,8 @@ experimental code should be included in oemof.experimental.
 
 from . import experimental
 from ._flow import Flow
-from ._investment_flow import InvestmentFlow
-from ._non_convex_flow import NonConvexFlow
-from ._non_convex_invest_flow import NonConvexInvestFlow
 
 __all__ = [
     "experimental",
     "Flow",
-    "InvestmentFlow",
-    "NonConvexFlow",
-    "NonConvexInvestFlow",
 ]
